# docs/swagger.yaml
basePath: /api
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  routes.Token:
    properties:
      access_token:
        type: string
      token_type:
        type: string
    type: object
...
  routes.UserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Secure digital banking platform with JWT authentication, TOTP 2FA,
    multi-currency wallets, atomic transfers, and admin monitoring
  title: SecureWallet API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/routes.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.Token'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      summary: Login user
      tags:
      - auth
...
  /backup:
    get:
      consumes:
      - application/json
      description: List all available database backups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List backups
      tags:
      - backup
  /backup/{filename}:
    get:
      consumes:
      - application/json
      description: Get information about a specific backup file
      parameters:
      - description: Backup filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get backup info
      tags:
      - backup
  /backup/config:
    get:
      consumes:
      - application/json
      description: Get backup configuration settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get backup config
      tags:
      - backup
  /backup/stats:
    get:
      consumes:
      - application/json
      description: Get backup statistics and metadata
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get backup stats
      tags:
      - backup
...
  /login-history:
    get:
      consumes:
      - application/json
      description: Get login history for the authenticated user
      parameters:
      - default: 50
        description: Number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get login history
      tags:
      - login-history
  /login-history/recent:
    get:
      consumes:
      - application/json
      description: Get recent login history for the authenticated user (default 10
        records)
      parameters:
      - default: 10
        description: Number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get recent login history
      tags:
      - login-history
  /security/alerts:
    get:
      consumes:
      - application/json
      description: Get security alerts with optional filtering by type, status, and
        limit (admin only)
      parameters:
      - description: Alert type filter
        in: query
        name: type
        type: string
      - description: Alert status filter
        in: query
        name: status
        type: string
      - description: Number of alerts to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get security alerts
      tags:
      - security
  /security/alerts/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a security alert (admin only)
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      - description: Alert status update data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update security alert status
      tags:
      - security
  /security/cleanup:
    post:
      consumes:
      - application/json
      description: Clean up old security monitoring data (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Cleanup security data
      tags:
      - security
  /security/idor/stats:
    get:
      consumes:
      - application/json
      description: Get IDOR (Insecure Direct Object Reference) detection statistics
        (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get IDOR detection stats
      tags:
      - security
  /security/users/{id}/reset-attempts:
    post:
      consumes:
      - application/json
      description: Reset security attempt counters for a specific user (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Reset user security attempts
      tags:
      - security
...
  /cron/execute/{job}:
    post:
      consumes:
      - application/json
      description: Execute a specific cron job manually
      parameters:
      - description: Job name
        in: path
        name: job
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Execute cron job
      tags:
      - cron
...
  /cron/remove:
    delete:
      consumes:
      - application/json
      description: Stop and remove all scheduled cron jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Remove cron jobs
      tags:
      - cron
...
  /cron/backup/stats:
    get:
      consumes:
      - application/json
      description: Get backup statistics from cron service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get backup stats from cron
      tags:
      - cron
securityDefinitions:
  BearerAuth:
    description: Enter the bearer token string with the "Bearer " prefix, e.g. "Bearer
      eyJhbGciOiJIUzI1NiIs..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"